Move model into package
